+++ lib/src/main/java/org/seuffert/panvpx/core/AbstractVpxEncoder.java
@@ -33,7 +33,7 @@ import org.seuffert.panvpx.ffi.vpx_rational;
  * </ol>
  *
  * <pre>{@code
- * try (Vp8Encoder encoder = new Vp8Encoder(VpxEncoderConfig.builder(640, 480).targetBitrateKbps(512).threads(2).build())) {
+ * try (Vp8Encoder encoder = new Vp8Encoder(VpxEncoderConfig.builder(VpxEncoderConfig.Codec.VP8, 640, 480).targetBitrateKbps(512).threads(2).build())) {
  *     long pts = 0;
  *     for (byte[] rawFrame : frameSource) {
  *         try (VpxImage image = VpxImage.fromByteArray(rawFrame, 640, 480)) {